+++ lib/session.js
@@ -1,4 +1,4 @@
-import { getWebsiteByUuid, getSessionByUuid, createSession } from 'lib/queries';
+import { getWebsiteByUuid, getSessionByUuid, createSession } from 'queries';
 import { getJsonBody, getClientInfo } from 'lib/request';
 import { uuid, isValidUuid, parseToken } from 'lib/crypto';
