+++ pages/api/account/index.js
@@ -1,4 +1,4 @@
-import { getAccountById, getAccountByUsername, updateAccount, createAccount } from 'lib/queries';
+import { getAccountById, getAccountByUsername, updateAccount, createAccount } from 'queries';
 import { useAuth } from 'lib/middleware';
 import { hashPassword } from 'lib/crypto';
 import { ok, unauthorized, methodNotAllowed, badRequest } from 'lib/response';
