+++ pages/api/website/[id]/active.js
@@ -1,7 +1,7 @@
-import { getActiveVisitors } from 'lib/queries';
 import { methodNotAllowed, ok, unauthorized } from 'lib/response';
 import { allowQuery } from 'lib/auth';
 import { useCors } from 'lib/middleware';
+import { getActiveVisitors } from 'queries';
 
 export default async (req, res) => {
   if (req.method === 'GET') {
