+++ pages/api/website/[id]/index.js
@@ -1,4 +1,4 @@
-import { deleteWebsite, getWebsiteById } from 'lib/queries';
+import { deleteWebsite, getWebsiteById } from 'queries';
 import { methodNotAllowed, ok, unauthorized } from 'lib/response';
 import { allowQuery } from 'lib/auth';
 import { useCors } from 'lib/middleware';
