+++ pages/api/website/[id]/stats.js
@@ -1,4 +1,4 @@
-import { getWebsiteStats } from 'lib/queries';
+import { getWebsiteStats } from 'queries';
 import { methodNotAllowed, ok, unauthorized } from 'lib/response';
 import { allowQuery } from 'lib/auth';
 import { useCors } from 'lib/middleware';
