+++ queries/admin/account/createAccount.js
@@ -1,0 +1,10 @@
+import { runQuery } from 'queries';
+import prisma from 'lib/db';
+
+export async function createAccount(data) {
+  return runQuery(
+    prisma.account.create({
+      data,
+    }),
+  );
+}
