+++ queries/admin/account/deleteAccount.js
@@ -1,0 +1,12 @@
+import { runQuery } from 'queries';
+import prisma from 'lib/db';
+
+export async function deleteAccount(user_id) {
+  return runQuery(
+    prisma.account.delete({
+      where: {
+        user_id,
+      },
+    }),
+  );
+}
