+++ queries/admin/account/updateAccount.js
@@ -1,0 +1,13 @@
+import { runQuery } from 'queries';
+import prisma from 'lib/db';
+
+export async function updateAccount(user_id, data) {
+  return runQuery(
+    prisma.account.update({
+      where: {
+        user_id,
+      },
+      data,
+    }),
+  );
+}
