+++ queries/admin/website/createWebsite.js
@@ -1,0 +1,17 @@
+import { runQuery } from 'queries';
+import prisma from 'lib/db';
+
+export async function createWebsite(user_id, data) {
+  return runQuery(
+    prisma.website.create({
+      data: {
+        account: {
+          connect: {
+            user_id,
+          },
+        },
+        ...data,
+      },
+    }),
+  );
+}
