+++ queries/admin/website/getUserWebsites.js
@@ -1,0 +1,15 @@
+import { runQuery } from 'queries';
+import prisma from 'lib/db';
+
+export async function getUserWebsites(user_id) {
+  return runQuery(
+    prisma.website.findMany({
+      where: {
+        user_id,
+      },
+      orderBy: {
+        name: 'asc',
+      },
+    }),
+  );
+}
