+++ queries/admin/website/getWebsiteByShareId.js
@@ -1,0 +1,12 @@
+import { runQuery } from 'queries';
+import prisma from 'lib/db';
+
+export async function getWebsiteByShareId(share_id) {
+  return runQuery(
+    prisma.website.findUnique({
+      where: {
+        share_id,
+      },
+    }),
+  );
+}
