+++ queries/admin/website/getWebsiteByUuid.js
@@ -1,0 +1,12 @@
+import { runQuery } from 'queries';
+import prisma from 'lib/db';
+
+export async function getWebsiteByUuid(website_uuid) {
+  return runQuery(
+    prisma.website.findUnique({
+      where: {
+        website_uuid,
+      },
+    }),
+  );
+}
