+++ queries/admin/website/resetWebsite.js
@@ -1,0 +1,6 @@
+import { runQuery } from 'queries';
+import prisma from 'lib/db';
+
+export async function resetWebsite(website_id) {
+  return runQuery(prisma.$queryRaw`delete from session where website_id=${website_id}`);
+}
