+++ queries/analytics/event/getEventMetrics.js
@@ -1,0 +1,28 @@
+import { getDateQuery, getFilterQuery, rawQuery } from 'queries';
+
+export function getEventMetrics(
+  website_id,
+  start_at,
+  end_at,
+  timezone = 'utc',
+  unit = 'day',
+  filters = {},
+) {
+  const params = [website_id, start_at, end_at];
+
+  return rawQuery(
+    `
+    select
+      event_value x,
+      ${getDateQuery('created_at', unit, timezone)} t,
+      count(*) y
+    from event
+    where website_id=$1
+    and created_at between $2 and $3
+    ${getFilterQuery('event', filters, params)}
+    group by 1, 2
+    order by 2
+    `,
+    params,
+  );
+}
