+++ queries/analytics/pageview/getPageviews.js
@@ -1,0 +1,19 @@
+import { runQuery } from 'queries';
+import prisma from 'lib/db';
+
+export async function getPageviews(websites, start_at) {
+  return runQuery(
+    prisma.pageview.findMany({
+      where: {
+        website: {
+          website_id: {
+            in: websites,
+          },
+        },
+        created_at: {
+          gte: start_at,
+        },
+      },
+    }),
+  );
+}
