+++ queries/analytics/session/getSessionByUuid.js
@@ -1,0 +1,12 @@
+import { runQuery } from 'queries';
+import prisma from 'lib/db';
+
+export async function getSessionByUuid(session_uuid) {
+  return runQuery(
+    prisma.session.findUnique({
+      where: {
+        session_uuid,
+      },
+    }),
+  );
+}
