+++ queries/index.js
@@ -1,0 +1,4 @@
+import * as admin from './admin';
+import * as analytics from './analytics';
+
+export default { admin, analytics };
